Sistema de análisis inteligente para operaciones rappi/requirements.txt
pandas>=2.0.0,<3.0.0
openpyxl>=3.1.0,<4.0.0
google-genai>=0.3.0,<1.0.0
reportlab>=4.0.0,<5.0.0